requirements.txt
diffusers==0.27.2
gradio==4.24.0
numpy==1.26.4
Pillow==10.2.0
